html/semantics/embedded-content-0/media-elements/loading-the-media-resource/038.html

<!DOCTYPE html>
<html><head>
	<title>'emptied' event fires after load()</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="../../../../../common/media.js"></script>
</head>
<body>
  <p><a href="http://www.w3.org/TR/html5/embedded-content-0.html#media-element-load-algorithm">Spec Reference</a></p>
  <video id="v" controls>
  </video>
  <div id="log"></div>
<script>
/* 4. If the media element's networkState is not set 
 * to NETWORK_EMPTY, then run these substeps:
 * 	1.Queue a task to fire a simple event named emptied 
 *	 at the media element... 
 */
var v = document.getElementById('v'),
	trigger_emptied = false;

v.src = getVideoURI("/media/movie_300") + "?" +
		new Date() + Math.random();

v.addEventListener("emptied", function() {
	trigger_emptied = true;
});

var t = async_test( "'emptied' event fired" );
	
v.load(); /* media element load algorithm */
	
setTimeout(function() {
	/* Wait 100ms before checking if event was fired */
	
	t.step(function() {
		/* First determine that networkState
		 * does NOT have value of NETWORK_EMPTY
		 */
		assert_not_equals(
			v.networkState,
			v.NETWORK_EMPTY,
			"videoElement.networkState should NOT be NETWORK_EMPTY"
		);
	});
		
	t.step(function() {
		/* Next check if the 'emptied' event was 
		 * triggered on v
		 */
		assert_true(
			trigger_emptied, "'emptied' was not fired."
		);
	});
	
	t.done();
}, 100);

</script>
	 
</body></html>
